Moved TFWeapon into the Items namespace.
