+++ TFModPlayer.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.Xna.Framework;
 using Terraria;
 using Terraria.ModLoader;
+using TerrariaFortress.Items;
 
 namespace TerrariaFortress
 {
